% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.VarBayesLR.R
\name{predict.VarBayesLR}
\alias{predict.VarBayesLR}
\title{predict.VarBayesLR}
\usage{
\method{predict}{VarBayesLR}(object, newdata, type = c("class", "prob")[2],
  ...)
}
\arguments{
\item{object}{Fitted model from \code{\link{VarBayesLR}}.}

\item{newdata}{A new input data frame.}

\item{type}{of prediction: "prop" for probabilities and "class" for
class labels.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
predictions depending on type
}
\description{
Make predictions using a fitted VarBayesLR model
}
\author{
Che Ngufor Ngufor.Che@mayo.edu
}

