% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SynData.R
\name{SynData}
\alias{SynData}
\title{SynData}
\usage{
SynData(n = 100, p = 10, type = c("two normal", "hyperplane")[1],
  class.prior = 0.3, noise = 0.01, ...)
}
\arguments{
\item{n}{sample size}

\item{p}{dimension}

\item{type}{type of artificial data.}

\item{class.prior}{proportion o positive examples}

\item{noise}{numeric constant in (0,1) for noise level}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A data frame
}
\description{
Generate two types of artificial binary classification data:
 two simple multivariate normal distributions with shifted centers
 and two multivariate normal distributions a noisy hyperplane.
}
\examples{
head(SynData())
}
\author{
Che Ngufor Ngufor.Che@mayo.edu
}

