% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{Utils}
\alias{Performance.measures}
\alias{Utils}
\alias{hprod}
\alias{lhs.form}
\alias{logdet}
\alias{mytanh}
\alias{opt.thresh}
\alias{rhs.form}
\alias{sigminus}
\alias{sigmoid}
\alias{sigplus}
\alias{tr}
\title{Utility functions.}
\usage{
sigmoid(x)

sigminus(x)

sigplus(x)

tr(x)

hprod(x, y)

logdet(x)

mytanh(x)

lhs.form(formula)

rhs.form(formula)

opt.thresh(pred, obs)

Performance.measures(pred, obs, threshold = NULL)
}
\arguments{
\item{x,y}{A numeric or character vector depending on the function}

\item{formula}{formula object}

\item{pred}{predicted probabilities}

\item{obs}{observed outcomes}

\item{threshold}{classification cut-off}
}
\value{
A numeric or character vector depending on the function
}
\description{
Useful functions
Some are used internally by other functions




}

