% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VarBayesLR.R
\name{VarBayesLR}
\alias{Mackay.Approximation}
\alias{VarBayesLR}
\alias{VarBayesLR.default}
\alias{VarBayesLR.formula}
\title{Variational Bayes Logistic Regression}
\usage{
VarBayesLR(X, ...)

\method{VarBayesLR}{default}(X, Y, max.iter = 100, tol = 1e-06,
  type.prob = c("Mckay", "logistic")[2], ...)

\method{VarBayesLR}{formula}(form, dat, max.iter = 100, tol = 1e-06,
  type.prob = c("Mckay", "logistic")[2], ...)

Mackay.Approximation(X, w, V)
}
\arguments{
\item{X}{data.frame/matrix of predictors.}

\item{Y}{binary (0/1) vector of response.}

\item{max.iter}{maximum number of iteration}

\item{tol}{convergence tolerance}

\item{type.prob}{type of approximation for posterior probabilities:
Mckay = Mckay approximation or logistic = sigmoid function the default}

\item{form}{formula}

\item{dat}{data.frame containing predictors and a binary (0/1) response
variable in formula.}

\item{w}{estimated predictor weights}

\item{V}{covariance matrix of weights}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
An object of class \code{VarBayesLR}; a list with items
\item{w}{weights}
\item{V}{weight covariance matrix}
\item{inV}{inverse of V}
\item{VarImp}{Variable Relevance: These are mean of gamma distribution: ratio of hyperparameters}
\item{L}{variational lower bound}
\item{prob}{fitted probabilities}
\item{thresh}{optimal classification threshold}
\item{type}{type of approximation for posterior probabilities}
}
\description{
Train Variational Bayesian Logistic Regression








}
\examples{
\dontrun{
set.seed(12345)
dat <- SynData()
ix = sample(nrow(dat), floor(nrow(dat)*0.75))
dat.trn = dat[ix, ]
dat.tst = dat[-ix, ]
form <- as.formula(paste("key ~ ", paste(names(dat)[!names(dat)\%in\%"key"], collapse = "+")))
mod <- VarBayesLR(form, dat.trn)
pred <- predict(mod, dat.tst)
perf <- Performance.measures(pred[,2], dat.tst$key, mod$thresh)
perf
}
}
\author{
Che Ngufor Ngufor.Che@mayo.edu
}
\references{
MacKay, David JC. "The evidence framework applied to classification networks."
Neural computation 4.5 (1992): 720-736.

Jaakkola, T., and Michael I. Jordan. "A variational approach to Bayesian logistic
regression models and their extensions." Sixth International Workshop on Artificial
Intelligence and Statistics. 1997.
}

